+++ rossum-agent/CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased] - YYYY-MM-DD
 - Fixed leaking Rossum API credentials across users' session [#41](https://github.com/stancld/rossum-mcp/pull/41)
+- Fixed leaking generated files across users' session [#42](https://github.com/stancld/rossum-mcp/pull/42)
 
 
 ## [0.1.2] - 2025-12-01